add LumiMask

+++ CatProducer/plugins/PileupWeightProducer.cc
@@ -58,9 +58,9 @@ PileupWeightProducer::PileupWeightProducer(const edm::ParameterSet& pset)
   puToken_ = pset.getParameter<edm::InputTag>("addPileupInfo");
 
   produces<int>("nTrueInteraction");
-  produces<double>("");
-  produces<double>("up");
-  produces<double>("dn");
+  produces<float>("");
+  produces<float>("up");
+  produces<float>("dn");
 
 }
 
@@ -70,9 +70,9 @@ void PileupWeightProducer::produce(edm::Event& event, const edm::EventSetup& eve
   event.getByLabel(puToken_, puHandle);
 
   std::auto_ptr<int> nTrueIntr(new int(-1));
-  std::auto_ptr<double> weight(new double(1.));
-  std::auto_ptr<double> weightUp(new double(1.));
-  std::auto_ptr<double> weightDn(new double(1.));
+  std::auto_ptr<float> weight(new float(1.));
+  std::auto_ptr<float> weightUp(new float(1.));
+  std::auto_ptr<float> weightDn(new float(1.));
   
   if ( !event.isRealData() and puHandle.isValid() )
   {
